requirements.txt
requests
tenacity
trafilatura
readability-lxml
pandas
pyarrow
python-dateutil
feedparser
beautifulsoup4
tqdm
langid
PyYAML
lxml
